requirements.txt
Django==1.10.5
django-extensions==1.7.6
djangorestframework==3.5.4
six==1.10.0
wsgiref==0.1.2
